fix eslint and vite warnings
