+++ src/pages/PlaygroundPage/PlaygroundPage.tsx
@@ -5,7 +5,6 @@ import InlineLoader from "@shared/ui/InlineLoader"
 import TraceSidePanel from "@shared/ui/TraceSidePanel"
 
 import {type AssemblyExecutionResult, executeAssemblyCode} from "@features/tasm/lib/executor.ts"
-import {useGlobalError} from "@shared/lib/errorContext"
 import StatusBadge, {type StatusType} from "@shared/ui/StatusBadge"
 import {useLineExecutionData, useTraceStepper} from "@features/txTrace/hooks"
 import {normalizeGas} from "@features/txTrace/lib/traceTx"
@@ -20,6 +19,8 @@ import {decodeCodeFromUrl} from "@app/pages/GodboltPage/urlCodeSharing.ts"
 import {ExecuteButton} from "@app/pages/PlaygroundPage/components/ExecuteButton.tsx"
 
 import {TUTORIAL_STEPS} from "@app/pages/PlaygroundPage/Tutorial.ts"
+
+import {useGlobalError} from "@shared/lib/useGlobalError.tsx"
 
 import styles from "./PlaygroundPage.module.css"
