+++ src/shared/ui/CodeEditor/hooks/useMonacoSetup.ts
@@ -2,7 +2,7 @@ import {useEffect, useState} from "react"
 import {useMonaco} from "@monaco-editor/react"
 import type * as monacoTypes from "monaco-editor"
 
-import {useTheme} from "@shared/lib/themeContext"
+import {useTheme} from "@shared/lib/useTheme"
 
 import {asmData} from "@features/tasm/lib"
